# Fix performance of `grafana_user`