@alexjfisher 2022-05-27
For reference, using the following puppet code and puppet apply to talk to a remote grafana server (in the same datacenter, but not as quick as localhost with no ssl!).

```puppet
grafana_user { range("testuser00","testuser99"):
  ensure           => present,                                                                                                                                                                                               
  grafana_url      => 'https://remotegrafana.example.com',                                                                                                                                                               
  grafana_user     => 'admin',                                                                                                                                                                                                            
  grafana_password => 'admin',                                                                                                                                                                                                            }
```
Before change...

Creating 100 users
Notice: Applied catalog in 126.45 seconds

Running Puppet again (ensuring those 100 users are still present)
Notice: Applied catalog in 224.31 seconds

Second run is slower as the average number of grafana users is > 100 for the entire duration, and not ~ 50.

After change...
1st puppet run
Notice: Applied catalog in 5.58 seconds
2nd puppet run
Notice: Applied catalog in 2.11 seconds

ie 100 times faster for this use case of just 100 users.